Create README.md
Leaflet.d3Layers is a collection of JavaScript functions used to easily draw D3 layers on top of a map built with Leaflet. 
The principal advantage in using D3 is the ability to manage large amount of data as vector SVG object mantaining very high performances: no more browser freeze or unresponsive maps!
Leaflet.d3Layers is under active development and it is really not recommended for your production environment.